requirements.txt
python-twitter==3.5
feedparser==6.0.8
urllib3==1.26.6